split auth and redirect handling back into two separate handlers
